+++ docker/circleci-cli/scripts/startup.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+
+# SYS: Set user permission for docker socket
+
+sudo chmod 666 /var/run/docker.sock
+
+# SYS: Make orb-cli stay running
+
+tail -f /dev/null