+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
- - Current development changes [ to be moved to release ]
+ - [ODP-1](https://github.com/serbiatech/ci-orb-docker-publisher/issues/1) Implement CircleCI CLI for local jobs execution. [ to be moved to release ]
 
+<!---
 ## [1.0.0] - YYYY-MM-DD
 ### Added
  - Initial Release
@@ -17,3 +18,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [1.0.0]: GITHUB TAG URL
+-->